// БР1.2/Персук Виктория/labs/lab1/src/common/enums.ts
export enum RoleName {
    Admin = 'Admin',
    User = 'User',
    Owner = 'Owner',
    Manager = 'Manager',
}

export enum RestaurantStatus {
    Pending = 'Pending',
    Verified = 'Verified',
    Rejected = 'Rejected',
}

export enum ReservationStatus {
    Confirmed = 'Confirmed',
    Cancelled = 'Cancelled',
    Pending = 'Pending',
}

export enum RatingEnum {
    One = 1,
    Two = 2,
    Three = 3,
    Four = 4,
    Five = 5,
}

export enum PriceCategory {
    Low = '$',
    Medium = '$$',
    High = '$$$',
}